fix is_ik_armature
- remove 'self' from args
- moved is_ik_armature, get_active_armature, get_deforming_armature,
  get_object_parent to utils.py
- __init__.py uses get_active_armature and is_ik_armature from utils
  (previous was algorithms that wasn't imported)

Issue #39
